python-s3-thumbnail/requirements.txt
boto3==1.16.17
botocore==1.19.17
jmespath==0.10.0
Pillow==8.0.1
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.26.1
